Add delete function and update related components

+++ animals/delete.php
@@ -1,16 +1,16 @@
 <?php
 include("../config/db.php");
 
-$id = $_GET['id'] ?? '';
-if($id){
-    // Optionally delete image file
+if(isset($_GET['id']) && $_GET['id'] != '') {
+    $id = $_GET['id'];
+
     $res = mysqli_query($conn, "SELECT Image FROM animals WHERE ID='$id'");
     $animal = mysqli_fetch_assoc($res);
-    if($animal && file_exists("../uploads/".$animal['Image'])){
+
+    if($animal && file_exists("../uploads/".$animal['Image'])) {
         unlink("../uploads/".$animal['Image']);
     }
 
-    // Delete from DB
     mysqli_query($conn, "DELETE FROM animals WHERE ID='$id'");
 }
